+++ cloudtrail.tf
@@ -1,7 +1,8 @@
 # CloudTrail
 resource "aws_cloudtrail" "global_Default" {
-  name                          = var.trail_name
-  s3_bucket_name                = local.bucket_name
+  count                         = local.create_cloudtrail
+  name                          = var.cloudtrail_name
+  s3_bucket_name                = local.cloudtrail_bucket_name
   is_multi_region_trail         = true
   include_global_service_events = true
   enable_log_file_validation    = true