+++ locals.tf
@@ -1,4 +1,6 @@
 locals {
+  # account alias 
+  create_account_alias = (var.create_account_alias == true ? 1 : 0)
   # cloudtrail
   create_cloudtrail        = (var.create_cloudtrail == true ? 1 : 0)
   create_cloudtrail_bucket = (var.create_cloudtrail_bucket == true && var.create_cloudtrail == true ? 1 : 0)