+++ iam_alias.tf
@@ -1,3 +1,5 @@
 resource "aws_iam_account_alias" "alias" {
-  account_alias = "${replace(lower(var.account_name),"_","-")}"
+  count         = local.create_account_alias
+  account_alias = replace(lower(var.account_alias), "_", "-")
 }
+